Splitting into header and source file. Also removing a bit of left-overs from MCP2515. This should have been done at the very beginning. It's impossible to work in a file that's so awful long and full of dead-code artefacts.
