+++ recipes-kernel/kernel-modules/x8h7/x8h7_can.c
@@ -2,6 +2,8 @@
  * X8H7 CAN driver
  */
 
+#include "x8h7_can.h"
+
 #include <linux/can/core.h>
 #include <linux/can/dev.h>
 #include <linux/can/led.h>
@@ -39,128 +41,6 @@
 #else
   #define DBG_CAN_STATE(i, s)
 #endif
-
-// Peripheral code
-#define X8H7_CAN1_PERIPH  0x03
-#define X8H7_CAN2_PERIPH  0x04
-// Op code
-#define X8H7_CAN_OC_INIT    0x10
-#define X8H7_CAN_OC_DEINIT  0x11
-#define X8H7_CAN_OC_BITTIM  0x12
-#define X8H7_CAN_OC_SEND    0x01
-#define X8H7_CAN_OC_RECV    0x01
-#define X8H7_CAN_OC_STS     0x40
-#define X8H7_CAN_OC_FLT     0x50
-
-#define X8H7_CAN_STS_INT_TX      0x01
-#define X8H7_CAN_STS_INT_RX      0x02
-#define X8H7_CAN_STS_INT_ERR     0x04
-/*
-MERRE: Message Error
-WAKIE: Wake-up Interrupt Enable
-ERRIE: Error Interrupt Enable bit (multiple sources in EFLG register
-TX2IE: Transmit Buffer 2 Empty
-RX1IE: Receive Buffer 1 Full I      questo non serve
-*/
-
-#define X8H7_CAN_STS_FLG_RX_OVR  0x01  // Receive Buffer Overflow
-#define X8H7_CAN_STS_FLG_TX_BO   0x02  // Bus-Off
-#define X8H7_CAN_STS_FLG_TX_EP   0x04  // Transmit Error-Passive
-#define X8H7_CAN_STS_FLG_RX_EP   0x08  // Receive Error-Passive
-#define X8H7_CAN_STS_FLG_TX_WAR  0x10  // Transmit Error Warning
-#define X8H7_CAN_STS_FLG_RX_WAR  0x20  // Receive Error Warning
-#define X8H7_CAN_STS_FLG_EWARN   0x40  // Error Warning
-#define X8H7_CAN_STS_FLG_TX_OVR  0x80  // Transmit Buffer Overflow
-
-#define X8H7_CAN_HEADER_SIZE        5
-#define X8H7_CAN_FRAME_MAX_DATA_LEN 8
-
-#define AFTER_SUSPEND_UP      1
-#define AFTER_SUSPEND_DOWN    2
-#define AFTER_SUSPEND_POWER   4
-#define AFTER_SUSPEND_RESTART 8
-
-#define X8H7_STD_FLT_MAX  128
-#define X8H7_EXT_FLT_MAX   64
-
-/**
- */
-union x8h7_can_init_message
-{
-  struct __attribute__((packed))
-  {
-    uint32_t baud_rate_prescaler;
-    uint32_t time_segment_1;
-    uint32_t time_segment_2;
-    uint32_t sync_jump_width;
-  } field;
-  uint8_t buf[sizeof(uint32_t) /* can_bitrate_Hz */ + sizeof(uint32_t) /* time_segment_1 */ + sizeof(uint32_t) /* time_segment_2 */ + sizeof(uint32_t) /* sync_jump_width */];
-};
-
-/**
- */
-union x8h7_can_bittiming_message
-{
-  struct __attribute__((packed))
-  {
-    uint32_t baud_rate_prescaler;
-    uint32_t time_segment_1;
-    uint32_t time_segment_2;
-    uint32_t sync_jump_width;
-  } field;
-  uint8_t buf[sizeof(uint32_t) /* can_bitrate_Hz */ + sizeof(uint32_t) /* time_segment_1 */ + sizeof(uint32_t) /* time_segment_2 */ + sizeof(uint32_t) /* sync_jump_width */];
-};
-
-/**
- */
-union x8h7_can_filter_message
-{
-  struct __attribute__((packed))
-  {
-    uint32_t idx;
-    uint32_t id;
-    uint32_t mask;
-  } field;
-  uint8_t buf[sizeof(uint32_t) /* idx */ + sizeof(uint32_t) /* id */ + sizeof(uint32_t) /* mask */];
-};
-
-/**
- */
-union x8h7_can_frame_message
-{
-  struct __attribute__((packed))
-  {
-    uint32_t id;
-    uint8_t  len;
-    uint8_t  data[X8H7_CAN_FRAME_MAX_DATA_LEN];
-  } field;
-  uint8_t buf[X8H7_CAN_HEADER_SIZE + X8H7_CAN_FRAME_MAX_DATA_LEN];
-};
-
-/**
- */
-struct x8h7_can_priv {
-  struct can_priv           can;
-  struct net_device        *net;
-  struct device            *dev;
-  int                       periph;
-
-  int                       tx_frame_rdy_cnt;
-  struct timer_list         tx_timer;
-  int                       tx_len;
-
-  struct workqueue_struct  *wq;
-  struct work_struct        tx_work;
-
-  int                       force_quit;
-  int                       after_suspend;
-  int                       restart_tx;
-
-  struct can_filter std_flt[X8H7_STD_FLT_MAX];
-  struct can_filter ext_flt[X8H7_EXT_FLT_MAX];
-
-  struct mutex        lock;
-};
 
 /**
  * Those parameters are valid for the STM32H7 CAN driver using
@@ -451,7 +331,6 @@ static int x8h7_can_open(struct net_device *net)
     return ret;
   }
 
-  priv->force_quit = 0;
   priv->tx_len  = 0;
   priv->tx_frame_rdy_cnt = 0;
 
@@ -590,7 +469,6 @@ static int x8h7_can_do_set_mode(struct net_device *net, enum can_mode mode)
   switch (mode) {
   case CAN_MODE_START:
     x8h7_can_clean(net);
-    /* We have to delay work since SPI I/O may sleep */
     priv->can.state = CAN_STATE_ERROR_ACTIVE;
     break;
   default:
@@ -605,8 +483,7 @@ static int x8h7_can_do_set_mode(struct net_device *net, enum can_mode mode)
 static int x8h7_can_do_get_berr_counter(const struct net_device *net,
                                         struct can_berr_counter *bec)
 {
-  //struct x8h7_can_priv *priv = netdev_priv(net);
-//@TODO: to be read from device
+  //@TODO: to be read from device
   bec->txerr = 0;
   bec->rxerr = 0;
 
@@ -832,8 +709,6 @@ static int x8h7_can_probe(struct platform_device *pdev)
     DBG_PRINT("fdcan_clk = %d", clock_freq);
   }
 
-//  init_waitqueue_head(&priv->wait);
-
   net = alloc_candev(sizeof(struct x8h7_can_priv), 1);
   if (!net) {
     return -ENOMEM;
@@ -852,8 +727,7 @@ static int x8h7_can_probe(struct platform_device *pdev)
   priv->can.do_get_berr_counter = x8h7_can_do_get_berr_counter;
   priv->can.ctrlmode_supported  = CAN_CTRLMODE_LOOPBACK      |
                                   CAN_CTRLMODE_LISTENONLY    |
-                                  CAN_CTRLMODE_3_SAMPLES     ;/*|
-                                  CAN_CTRLMODE_BERR_REPORTING;*/
+                                  CAN_CTRLMODE_3_SAMPLES     ;
   priv->net = net;
 
   platform_set_drvdata(pdev, priv);
@@ -923,7 +797,6 @@ MODULE_DEVICE_TABLE(platform, x8h7_can_id_table);
 static struct platform_driver x8h7_can_driver = {
   .driver = {
     .name           = DRIVER_NAME,
-    //.pm             = &x8h7_can_pm_ops,
     .of_match_table = x8h7_can_of_match,
   },
   .probe    = x8h7_can_probe,
